Unit11/secure_shop/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-CORS==4.0.0
python-decouple==3.8
SQLAlchemy==2.0.23
pydantic==2.5.2
email-validator==2.1.0.post1

# Security
bcrypt==4.0.1
PyJWT==2.8.0
pyotp==2.9.0

# Database and Caching
redis==5.0.1
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Production
gunicorn==21.2.0
